+++ src/main/java/tk/parking/app/service/ParkingService.java
@@ -3,5 +3,5 @@ package tk.parking.app.service;
 import tk.parking.app.http.request.ParkingDTO;
 
 public interface ParkingService {
-    Long createParking(final ParkingDTO parkingDTO);
+    Integer createParking(final ParkingDTO parkingDTO);
 }
